Task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -8,7 +8,7 @@
  * Return: the number of nodes
  */
 size_t print_dlistint(const dlistint_t *h) {
-	size_t nodes;
+	int nodes;
 
 	nodes = 0;
 	if (h == NULL) {
